requirements.txt
azure-storage-blob==12.12.0
blis==0.7.8
catalogue==2.0.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
confection==0.0.3
cymem==2.0.6
Cython==0.29.28
certifi==2022.6.15
cffi==1.15.0
cryptography==37.0.2
docx2txt==0.8
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
gensim==4.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
lxml==4.9.1
MarkupSafe==2.1.1
murmurhash==1.0.8
faiss-cpu==1.7.2
filelock==3.7.1
huggingface-hub==0.8.1
importlib-metadata==4.11.4
isodate==0.6.1
msrest==0.7.1
nltk==3.7
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pathy==0.6.2
preshed==3.0.7
pydantic==1.9.2
PyMuPDF==1.20.2
pyparsing==3.0.9
PyPDF2==2.11.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.4
oauthlib==3.2.0
pdfminer.six
Pillow==9.1.1
pycparser==2.21
pycryptodome==3.15.0
PyYAML==6.0
regex==2022.9.13
requests==2.28.0
requests-oauthlib==1.3.1
scikit-learn==1.1.2
scipy==1.9.2
sentence-transformers==2.2.1
sentencepiece==0.1.96
six==1.16.0
selenium==4.8.2
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0
torchvision==0.12.0
transformers==4.20.1
XlsxWriter==3.0.3
zipp==3.8.0
pdfplumber
word2number
openpyxl
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0.tar.gz#egg=en_core_web_sm
flask_cors
sparse-dot-topn
sqlalchemy-orm==1.2.3
azure-data-tables
gunicorn
uvicorn
smart-open==5.2.1
srsly==2.4.4
thinc==8.1.4
threadpoolctl==3.1.0
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
Werkzeug==2.2.2
docx2pdf
pyodbc
sqlalchemy
https://github.com/explosion/spacy-models/releases/download/nl_core_news_sm-3.4.0/nl_core_news_sm-3.4.0.tar.gz
pyhtml2pdf
python-dotenv==1.0.0
Flask-JWT-Extended==4.4.4
